finished merging oidc and oauth2 [#523]
