+++ .secrets.baseline
@@ -124,7 +124,7 @@
         "filename": "core/auth/apis.go",
         "hashed_secret": "4d55af37dbbb6a42088d917caa1ca25428ec42c9",
         "is_verified": false,
-        "line_number": 1937
+        "line_number": 1939
       }
     ],
     "core/auth/auth.go": [
@@ -193,7 +193,7 @@
         "filename": "core/auth/auth_type_oauth2.go",
         "hashed_secret": "f8dabcd142517fb16418a3a2790f3da31392537a",
         "is_verified": false,
-        "line_number": 130
+        "line_number": 122
       }
     ],
     "core/auth/auth_type_oidc.go": [
@@ -202,7 +202,7 @@
         "filename": "core/auth/auth_type_oidc.go",
         "hashed_secret": "f8dabcd142517fb16418a3a2790f3da31392537a",
         "is_verified": false,
-        "line_number": 122
+        "line_number": 113
       }
     ],
     "driven/emailer/adapter.go": [
@@ -288,5 +288,5 @@
       }
     ]
   },
-  "generated_at": "2022-08-23T22:08:15Z"
+  "generated_at": "2022-08-24T20:24:53Z"
 }
